Refactor: Cambio de localización de la navegacion
